add more functional interfaces rename some functional interfaces add applicable to some interfaces

+++ src/de/tinycodecrank/functions/float_/applicable/FloatApplicableRight.java
@@ -3,5 +3,5 @@ package de.tinycodecrank.functions.float_.applicable;
 @FunctionalInterface
 public interface FloatApplicableRight<R>
 {
-	R aptLast(int z);
+	R aptLast(float z);
 }
